+++ ComplexPortfolio.Module/HelpClasses/ProportionData.cs
@@ -6,14 +6,14 @@ using System.Threading.Tasks;
 
 namespace ComplexPortfolio.Module.HelpClasses {
     public class ProportionData {
-        public ProportionData(string _ticker,decimal _price, int _currenAmount,double _targetProportion) {
+        public ProportionData(string _ticker,double _price, int _currenAmount,double _targetProportion) {
             Ticker = _ticker;
             Price = _price;
             CurrentAmount = _currenAmount;
             TargetProportion = _targetProportion;
         }
         public string Ticker { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public double TargetProportion { get; set; }
         public int CurrentAmount { get; set; }
     }
